requirements.txt
# Core ML
torch>=2.5.0
torchvision>=0.20.0
# Install diffusers from git for FLUX.2 Klein + LTX-2 support:
# pip install git+https://github.com/huggingface/diffusers.git
diffusers>=0.36.0
transformers>=5.0.0
accelerate>=1.0.0
safetensors>=0.4.0

# API
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Redis
redis[hiredis]>=5.0.0

# Storage
supabase>=2.7.0
httpx>=0.27.0

# Media
Pillow>=10.0
imageio>=2.33
imageio-ffmpeg>=0.4.0

# HuggingFace
huggingface_hub>=1.5.0

# LTX-2 native pipelines (for IC-LoRA pose transfer)
# pip install git+https://github.com/Lightricks/LTX-2.git
